Check platform env vars
